Added event for PasswordEntry changes
