+++ Boba.PasswordManagerUnitTests/PasswordLibraryUnitTests.cs
@@ -1,0 +1,72 @@
+﻿// Boba Password Manager (https://github.com/an-prata/Boba)
+// Copyright (c) 2021 Evan Overman (https://github.com/an-prata)
+// Licensed under the MIT License.
+
+using System;
+using System.Text;
+using System.Collections.Generic;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Boba.PasswordManager;
+
+namespace Boba.PasswordManagerTests
+{
+	[TestClass]
+	class PasswordLibraryUnitTests
+	{
+		//private readonly string[] Applications = { "google.com", "docs.microsoft.com", "github.com", "stackoverflow.com" };
+		//private readonly string[] Usernames = { "jake@gmail.com", "jake", "jakemiester1234", "jakesmith@outlook.com" };
+		//private readonly string[] Passwords = { "jakesSuperStrongPassword", "Password12345", "TheBatmanEmoji", "#sYfg@sh&*gskks%$#17291$sg%" };
+
+		private readonly string[,] Credentials = {
+			{ "google.com", "jake@gmail.com", "jakesSuperStrongPassword" },
+			{ "docs.microsoft.com", "jake", "Password12345" },
+			{ "github.com", "jakemiester1234", "TheBatmanEmoji"},
+			{ "stackoverflow.com", "jakesmith@outlook.com", "#sYfg@sh&*gskks%$#17291$sg%"}
+		};
+
+		private readonly string[,] SortedCredentials = {
+			{ "docs.microsoft.com", "jake", "Password12345" },
+			{ "github.com", "jakemiester1234", "TheBatmanEmoji"},
+			{ "google.com", "jake@gmail.com", "jakesSuperStrongPassword" },
+			{ "stackoverflow.com", "jakesmith@outlook.com", "#sYfg@sh&*gskks%$#17291$sg%"}
+		};
+
+		private int PasswordEntriesChangedEventRaises { get; set; }
+
+		[TestMethod]
+		public void AlphabetizationTest()
+		{
+			PasswordLibrary passwordLibrary = new PasswordLibrary("test library", new List<PasswordEntry>());
+			PasswordLibrary preSortedPasswordLibrary = new PasswordLibrary("test library", new List<PasswordEntry>());
+
+			for (int i = 0; i < Credentials.GetLength(1); i++) 
+				passwordLibrary.NewEntry(Encoding.UTF8.GetBytes(Credentials[i, 2]), Credentials[i, 1], Credentials[i, 0]);
+
+			for (int i = 0; i < Credentials.GetLength(1); i++) 
+				preSortedPasswordLibrary.NewEntry(Encoding.UTF8.GetBytes(SortedCredentials[i, 2]), SortedCredentials[i, 1], SortedCredentials[i, 0]);
+
+			for (int i = 0; i < passwordLibrary.PasswordEntries.Count; i++) 
+				Assert.AreEqual(passwordLibrary.PasswordEntries[i].Application, preSortedPasswordLibrary.PasswordEntries[i].Application);
+		}
+
+		[TestMethod]
+		public void PasswordEntriesChangedEventTest()
+		{
+			PasswordEntriesChangedEventRaises = 0;
+			int expectedPasswordEntriesChangedRaises = 3;
+			PasswordLibrary passwordLibrary = new PasswordLibrary("test library", new List<PasswordEntry>());
+			passwordLibrary.PasswordEntriesChanged += delegate(object sender, EventArgs e) { ++PasswordEntriesChangedEventRaises; };
+
+			passwordLibrary.NewEntry(Encoding.UTF8.GetBytes(Credentials[0, 2]), Credentials[0, 1], Credentials[0, 0]);
+			passwordLibrary.PasswordEntries = new List<PasswordEntry>();
+
+			passwordLibrary.PasswordEntries = new List<PasswordEntry> 
+			{ 
+				new PasswordEntry(Encoding.UTF8.GetBytes(Credentials[1, 2]), Credentials[1, 1], Credentials[1, 0]), 
+				new PasswordEntry(Encoding.UTF8.GetBytes(Credentials[2, 2]), Credentials[2, 1], Credentials[2, 0]) 
+			};
+
+			Assert.AreEqual(expectedPasswordEntriesChangedRaises, PasswordEntriesChangedEventRaises);
+		}
+	}
+}
